backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-Cors==4.0.0
werkzeug==3.0.1
gunicorn==21.2.0

# Audio Processing & Music
numpy==1.26.2
# [중요] scipy 충돌 방지를 위해 librosa 버전 올림
librosa>=0.10.2
soundfile==0.12.1
pretty_midi==0.2.10
music21==9.1.0
demucs==4.0.1

# Deep Learning (PyTorch CPU version)
torch==2.1.2
torchaudio==2.1.2

# Utilities
tqdm==4.66.1
python-dotenv>=1.0.0
requests==2.31.0

# Google Cloud Service (GCS)
google-cloud-storage==2.14.0
